+++ tests/letstest/scripts/test_leauto_upgrades.sh
@@ -1,7 +1,7 @@
 #!/bin/bash -xe
 set -o pipefail
 
-# $OS_TYPE $PUBLIC_IP $PRIVATE_IP $PUBLIC_HOSTNAME $BOULDER_URL
+# $OS_TYPE $PUBLIC_IP $PRIVATE_IP $PUBLIC_HOSTNAME
 # are dynamically set at execution
 
 cd letsencrypt
@@ -43,9 +43,11 @@ fi
 # directory to be served.
 MY_TEMP_DIR=$(mktemp -d)
 PORT_FILE="$MY_TEMP_DIR/port"
+LOG_FILE="$MY_TEMP_DIR/log"
 SERVER_PATH=$("$PYTHON_NAME" tools/readlink.py tools/simple_http_server.py)
 cd "$MY_TEMP_DIR"
-"$PYTHON_NAME" "$SERVER_PATH" 0 > $PORT_FILE &
+# We set PYTHONUNBUFFERED to disable buffering of output to LOG_FILE
+PYTHONUNBUFFERED=1 "$PYTHON_NAME" "$SERVER_PATH" 0 > $PORT_FILE 2> "$LOG_FILE" &
 SERVER_PID=$!
 trap 'kill "$SERVER_PID" && rm -rf "$MY_TEMP_DIR"' EXIT
 cd ~-
@@ -105,15 +107,10 @@ if ./letsencrypt-auto -v --debug --version | grep "WARNING: couldn't find Python
     exit 1
 fi
 
-# On systems like Debian where certbot-auto is deprecated, we expect it to
-# leave existing Certbot installations unmodified so we check for the same
-# version that was initially installed below.  Once certbot-auto is deprecated
-# on RHEL systems, we can unconditionally check for INITIAL_VERSION.
-if [ -f /etc/debian_version ]; then
-    EXPECTED_VERSION="$INITIAL_VERSION"
-else
-    EXPECTED_VERSION=$(grep -m1 LE_AUTO_VERSION certbot-auto | cut -d\" -f2)
-fi
+# Since certbot-auto is deprecated, we expect it to leave existing Certbot
+# installations unmodified so we check for the same version that was initially
+# installed below.
+EXPECTED_VERSION="$INITIAL_VERSION"
 
 if ! /opt/eff.org/certbot/venv/bin/letsencrypt --version 2>&1 | tail -n1 | grep "^certbot $EXPECTED_VERSION$" ; then
     echo unexpected certbot version found
@@ -125,21 +122,47 @@ if ! diff letsencrypt-auto letsencrypt-auto-source/letsencrypt-auto ; then
     exit 1
 fi
 
-if [ "$RUN_RHEL6_TESTS" = 1 ]; then
-    # Add the SCL python release to PATH in order to resolve python3 command
-    PATH="/opt/rh/rh-python36/root/usr/bin:$PATH"
-    if ! command -v python3; then
-        echo "Python3 wasn't properly installed"
-        exit 1
-    fi
-    if [ "$(/opt/eff.org/certbot/venv/bin/python -V 2>&1 | cut -d" " -f 2 | cut -d. -f1)" != 3 ]; then
-        echo "Python3 wasn't used in venv!"
-        exit 1
-    fi
+# Now let's test if letsencrypt-auto still tries to upgrade to a new version.
+# Regardless of the OS, versions of the script with development version numbers
+# ending in .dev0 will not upgrade. See
+# https://github.com/certbot/certbot/blob/bdfb9f19c4086a60ef010d2431768850c26d838a/certbot-auto#L1947-L1948.
+# In order to test the process of different OSes setting NO_SELF_UPGRADE as
+# part of the script's deprecation, we make use of the fact that
+# letsencrypt-auto should still attempt to fetch the version number from PyPI
+# even if it has a development version number unless NO_SELF_UPGRADE is set in
+# which case all of that logic should be skipped.
+#
+# First we make a copy of the current server logs.
+PREVIOUS_LOG_FILE="$MY_TEMP_DIR/previous-log"
+cp "$LOG_FILE" "$PREVIOUS_LOG_FILE"
 
-    if [ "$("$PYTHON_NAME" tools/readlink.py $OLD_VENV_PATH)" != "/opt/eff.org/certbot/venv" ]; then
-        echo symlink from old venv path not properly created!
+# Next we run letsencrypt-auto and make sure there were no problems checking
+# for updates, the Certbot install still works, the version number is what
+# we expect, and it prints a message about not receiving updates.
+if ./letsencrypt-auto -v --debug --version | grep "WARNING: couldn't find Python" ; then
+    echo "Had problems checking for updates!"
+    exit 1
+fi
+if ! ./letsencrypt-auto -v --debug --version 2>&1 | tail -n1 | grep "^certbot $EXPECTED_VERSION$" ; then
+    echo unexpected certbot version found
+    exit 1
+fi
+if ! ./letsencrypt-auto -v --debug --version 2>&1 | grep "will no longer receive updates" ; then
+    echo script did not print warning about not receiving updates!
+    exit 1
+fi
+
+# Finally, we check if our local server received more requests. Over time,
+# we'll move more and more OSes into this case until it this is the expected
+# behavior on all systems.
+if [ -f /etc/issue ] && grep -iq "Amazon Linux" /etc/issue; then
+    if ! diff "$LOG_FILE" "$PREVIOUS_LOG_FILE" ; then
+        echo our local server received unexpected requests
+        exit 1
+    fi
+else
+    if diff "$LOG_FILE" "$PREVIOUS_LOG_FILE" ; then
+        echo our local server did not receive the requests we expected
         exit 1
     fi
 fi
-echo upgrade appeared to be successful